Moved listener and handler inside TCPManager
